move mockprocessmap to separate file
